+++ tests/data/modules/crudmodulefolder/admin/Module.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace luyatests\data\modules\crudmodulefolder\admin;
+
+use luya\base\Module;
+
+class Module extends Module
+{
+	
+}